Add files via upload
This project is a classic machine learning exercise where I tried to learn which passengers of the Titanic survived using the Logistic Regression algorithm
This repository contains the datasets and the submission results. You can download it and use it as you wish and use my submission dataset as an example in order to learn.